Add check to make sure dist exists before trying to clear it

+++ scripts/clean.ps1
@@ -1,2 +1,5 @@
-Remove-Item "./dist/" -Recurse
+if (Test-Path "./dist") {
+    Remove-Item "./dist/" -Recurse
+}
+
 Get-ChildItem "./src/" -Include "*.ngfactory.ts", "*.ngsummary.json" -Recurse | Remove-Item
